+++ internal/test-koin/src/commonMain/kotlin/com/github/panpf/sketch/test/utils/Koins.kt
@@ -1,0 +1,20 @@
+package com.github.panpf.sketch.test.utils
+
+import com.github.panpf.sketch.Sketch
+import org.koin.core.context.startKoin
+import org.koin.dsl.module
+
+object Koins {
+
+    init {
+        startKoin {
+            modules(module {
+                single { Sketch.Builder(getTestContext()).build() }
+            })
+        }
+    }
+
+    fun initial() {
+
+    }
+}